PLATFORM-6985 | Apply Hydra fixes to Cargo

+++ CargoHooks.php
@@ -156,7 +156,7 @@ class CargoHooks {
 		// Get all the "main" tables that this page is contained in.
 		$dbw = wfGetDB( DB_MASTER );
 		$cdb = CargoUtils::getDB();
-		$cdb->begin();
+		$cdb->startAtomic( __METHOD__ );
 		$cdbPageIDCheck = [ $cdb->addIdentifierQuotes( '_pageID' ) => $pageID ];
 
 		$res = $dbw->select( 'cargo_pages', 'table_name', [ 'page_id' => $pageID ] );
@@ -171,16 +171,18 @@ class CargoHooks {
 			// First, delete from the "field" tables.
 			$fieldTablesValue = $dbw->selectField( 'cargo_tables', 'field_tables', [ 'main_table' => $curMainTable ] );
 			$fieldTableNames = unserialize( $fieldTablesValue );
-			foreach ( $fieldTableNames as $curFieldTable ) {
-				// Thankfully, the MW DB API already provides a
-				// nice method for deleting based on a join.
-				$cdb->deleteJoin(
-					$curFieldTable,
-					$curMainTable,
-					$cdb->addIdentifierQuotes( '_rowID' ),
-					$cdb->addIdentifierQuotes( '_ID' ),
-					$cdbPageIDCheck
-				);
+			if ( is_array( $fieldTableNames ) ) {
+				foreach ( $fieldTableNames as $curFieldTable ) {
+					// Thankfully, the MW DB API already provides a
+					// nice method for deleting based on a join.
+					$cdb->deleteJoin(
+						$curFieldTable,
+						$curMainTable,
+						$cdb->addIdentifierQuotes( '_rowID' ),
+						$cdb->addIdentifierQuotes( '_ID' ),
+						$cdbPageIDCheck
+					);
+				}
 			}
 
 			// Delete from the "files" helper table, if it exists.
@@ -206,7 +208,7 @@ class CargoHooks {
 		CargoBackLinks::managePageDeletion( $pageID );
 
 		// End transaction and apply DB changes.
-		$cdb->commit();
+		$cdb->endAtomic( __METHOD__ );
 	}
 
 	public static function deletePageFromSpecialTable( $pageID, $specialTableName ) {
@@ -380,7 +382,8 @@ class CargoHooks {
 		}
 		$dbw = wfGetDB( DB_MASTER );
 		$cdb = CargoUtils::getDB();
-		$cdb->begin();
+
+		$cdb->startAtomic( __METHOD__ );
 
 		$res = $dbw->select( 'cargo_pages', 'table_name', [ 'page_id' => $pageid ] );
 		foreach ( $res as $row ) {
@@ -415,12 +418,13 @@ class CargoHooks {
 		}
 
 		// End transaction and apply DB changes.
-		$cdb->commit();
+		$cdb->endAtomic( __METHOD__ );
 
 		// Save data for the original page (now a redirect).
 		if ( $redirid != 0 ) {
+			$oldTitle = Title::newFromLinkTarget( $old );
 			$useReplacementTable = $cdb->tableExists( '_pageData__NEXT' );
-			CargoPageData::storeValuesForPage( $title, $useReplacementTable );
+			CargoPageData::storeValuesForPage( $oldTitle, $useReplacementTable );
 		}
 	}
